requirements.txt
Django==2.1.3
django-crispy-forms==1.7.2
pytz==2018.7
